Resolve lint warning and errors
Previously committing without the -n flag would fail. To resolve this
issue, I've resolved all of the warnings and errors that the
application produced with the current set of linting rules.

In most cases I've disabled the problematic rule for the file in
question--I did not have time nor the understanding to add prop types
for N amount of components.

Some of the linting errors were odd. For instance, some components used
== instead of === which is something you rarely see. Some of the CSS
styles also had oddities.