Add initial day calendar view
In these changes we are importing the Fullcalendar component we use in
Varaamo. I attempted to make the addition with minimal required
changes. Some polishing is still needed, but I should be able to make
those when I am implementing the week view.

I also shuffled the version we are using of some libraries, because
Fullcalendar was dependent on these.