Install dependencies once
The current build script tooling doesn't offer good supprt for ignoring
dependencies that are not required for production. For this reason we
always end up installing all the dependencies. To save build time, I'm
removing future proofing which would have made it easier to install
fewer dependencies for production in the future.