Stop rendering event duration in week view
The event element within the week view is so limited in space that it
can't fit the duration information for the event. That's why we've
removed that information entirely.

Unfortunately this resulted a lot of refactoring. In order to reliably
get the current viewType within handleEventRender, we need to source
the viewType from the calendarApi directly. This meant that the
viewType in state would have been a second source of truth, which we do
do not want. Removing viewType from the state created a need for other
small changes.

Fullcalendar does not provide a method for hooking into view change
events. To mock this functionality, we've had to create custom buttons
which allows us to define onClick handlers where we can hook into view
changes. After these changes we've also been able to decrease the
amount of prop and state syncing that happens in the component. This
has resolved the need to include a shouldComponentUpdate hack which was
used to stop the component from updating when it would have resulted in
an errorenous state.

Previously we hooked into viewType changes by syncing the viewState in
the calendarApi on each component update. This resulted in situations
where the current viewType was updated after renders had been
completed.